Add Initiatives tab
